+++ plugins/caspar/app/components/LibraryListItem/index.jsx
@@ -32,6 +32,7 @@ const ITEM_CONSTRUCTORS = [
   {
     if: item => [asset.type.still, asset.type.video, asset.type.audio].includes(item.type),
     fn: item => {
+      console.log('Item', item)
       return {
         type: 'bridge.caspar.media',
         data: {
@@ -39,6 +40,7 @@ const ITEM_CONSTRUCTORS = [
           caspar: {
             server: item?._filter?.serverId,
             target: item.name,
+            frameRate: asset.frameRateFractionToDecimalRounded(item?.framerate || asset.DEFAULT_FRAMERATE_FRACTION),
             ...(DEFAULT_VALUES[item.type] || {})
           },
           duration: asset.calculateDurationMs(item)